feat: try to process event only after expected finalization
It drastically reduces number of attempts. For example, relayer tried to process event every 10 seconds immediately after getting an event and on Base and it resulted in 6 (reqs per minute) * 18 (avg finalization time on Eth in minutes) = 108 reqs. New method tries to process only after waiting for around 1000 seconds (around 16.5 minutes and this number is different for each chain and it can be configured in `config.toml`)